+++ Loader/src/utils/helpers.cpp
@@ -471,20 +471,20 @@ namespace toad
 		draw->PathClear();// window->DrawList->PathClear();
 
 		int num_segments = 30;
-		int start = abs(ImSin(g.Time * 1.8f) * (num_segments - 5));
+		float start = abs(ImSin((float)g.Time * 1.8f) * (num_segments - 5));
 
-		const float a_min = IM_PI * 2.0f * ((float)start) / (float)num_segments;
+		const float a_min = IM_PI * 2.0f * start / (float)num_segments;
 		const float a_max = IM_PI * 2.0f * ((float)num_segments - 3) / (float)num_segments;
 
 		const auto centre = ImVec2(pos.x + radius, pos.y + radius + style.FramePadding.y);
 
 		for (int i = 0; i < num_segments; i++) {
 			const float a = a_min + ((float)i / (float)num_segments) * (a_max - a_min);
-			draw->PathLineTo(ImVec2(centre.x + ImCos(a + g.Time * 8) * radius,
-				centre.y + ImSin(a + g.Time * 8) * radius));
+			draw->PathLineTo(ImVec2(centre.x + (float)ImCos(a + (float)g.Time * 8) * radius,
+				centre.y + (float)ImSin(a + (float)g.Time * 8) * radius));
 		}
 
-		draw->PathStroke(color, false, thickness);
+		draw->PathStroke(color, false, (float)thickness);
 		return true;
 	}
 }
