+++ ToadClient/src/Toad/Modules/clicker/clicker_base.cpp
@@ -8,7 +8,7 @@ namespace toadll
 	{
 		auto& rand = get_rand();
 
-		rand.inconsistency_delay = std::lerp(rand.inconsistency_delay, 0, m_pTick * 1.5f);
+		rand.inconsistency_delay = std::lerp(rand.inconsistency_delay, 0.f, m_pTick * 1.5f);
 
 		const int rand_100 = rand_int(0, 100);
 
@@ -18,7 +18,7 @@ namespace toadll
 				if (i.chance >= rand_100)
 					i.start = true;
 				else
-					i.frequency_counter -= static_cast<float>(i.frequency_counter) * (static_cast<float>(i.frequency_counter) / static_cast<float>(i.frequency) * 0.75f);
+					i.frequency_counter -= i.frequency_counter * (int)((float)i.frequency_counter / (float)i.frequency * 0.75f);
 			}
 
 		for (auto& i2 : rand.inconsistencies2)
@@ -27,7 +27,7 @@ namespace toadll
 				if (i2.chance <= rand_100)
 					i2.start = true;
 				else
-					i2.frequency_counter -= static_cast<float>(i2.frequency_counter) * (static_cast<float>(i2.frequency_counter) / static_cast<float>(i2.frequency) * 0.75f);
+					i2.frequency_counter -= i2.frequency_counter * (int)((float)i2.frequency_counter / (float)i2.frequency * 0.75f);
 			}
 
 		for (auto& b : rand.boosts)
@@ -45,7 +45,7 @@ namespace toadll
 				{
 					if (b_other.id == b.id)
 						continue;
-					b_other.frequency_counter -= static_cast<float>(b_other.frequency_counter) * (static_cast<float>(b_other.frequency_counter) / static_cast<float>(b_other.frequency) * 0.75f);
+					b_other.frequency_counter -= b_other.frequency_counter * (int)((float)b_other.frequency_counter / (float)b_other.frequency * 0.75f);
 					b_other.paused = true;
 				}
 			}
